+++ tiptapy/__init__.py
@@ -1,31 +1,35 @@
+import json
 import os
 import sys
-import json
 from html import escape
-from jinja2 import FileSystemLoader, Environment, select_autoescape
 from typing import Dict
+
+from jinja2 import Environment, FileSystemLoader, select_autoescape
+
 from .image import url2mime
-from .macros import (make_img_src, build_link_handler,
-                     get_audio_player_block, get_doc_block)
+from .macros import (
+    build_link_handler,
+    get_audio_player_block,
+    get_doc_block,
+    make_img_src,
+)
 
-
-__version__ = '0.17.0'
+__version__ = "0.18.0"
 
 renderers: Dict = {}
 
 
 def init_env(path, config):
-    env = Environment(loader=FileSystemLoader(path),
-                      autoescape=select_autoescape(
-                          enabled_extensions=('html')))
+    env = Environment(
+        loader=FileSystemLoader(path),
+        autoescape=select_autoescape(enabled_extensions=("html")),
+    )
     # https://stackoverflow.com/a/6038550
-    env.globals['url2mime'] = url2mime
-    env.globals['make_img_src'] = make_img_src
-    env.globals['handle_links'] = build_link_handler(config)
-    # Cause jinja2 `e` filter is not exactly same as html.escape
-    env.globals['escape'] = escape
-    env.globals['get_audio_player_block'] = get_audio_player_block
-    env.globals['get_doc_block'] = get_doc_block
+    env.globals["url2mime"] = url2mime
+    env.globals["make_img_src"] = make_img_src
+    env.globals["handle_links"] = build_link_handler(config)
+    env.globals["get_audio_player_block"] = get_audio_player_block
+    env.globals["get_doc_block"] = get_doc_block
 
     return env
 
@@ -39,11 +43,26 @@ def _get_abs_template_path(path_str):
     return os.path.join(pkg_dir, path_str)
 
 
-class BaseDoc:
+def escape_values_recursive(node):
+    skip_key = "html"  # Skip escaping html values in embeds
+    if isinstance(node, dict):
+        for k, v in node.items():
+            if k != skip_key:
+                node[k] = escape_values_recursive(v)
+    elif isinstance(node, list):
+        for i, v in enumerate(node):
+            node[i] = escape_values_recursive(v)
+    elif isinstance(node, str):
+        return escape(node)
+    return node
 
-    doc_type = 'doc'
-    templates_path = (_get_abs_template_path('templates'),
-                      _get_abs_template_path('templates/extras'))
+
+class BaseDoc:
+    doc_type = "doc"
+    templates_path = (
+        _get_abs_template_path("templates"),
+        _get_abs_template_path("templates/extras"),
+    )
     locked = False
     # `locked` helps in templates determine to show/hide in anonymous views
     # Useful in case where code is referring same template for both protected
@@ -51,10 +70,11 @@ class BaseDoc:
 
     def __init__(self, config):
         environ = init_env(self.templates_path, config)
-        self.t = environ.get_template(f'{self.doc_type}.html')
-        self.t.environment.globals['locked'] = self.locked
+        self.t = environ.get_template(f"{self.doc_type}.html")
+        self.t.environment.globals["locked"] = self.locked
 
     def render(self, in_data):
         in_data = in_data if isinstance(in_data, dict) else json.loads(in_data)
         node = in_data if isinstance(in_data, dict) else json.loads(in_data)
+        node = escape_values_recursive(node)
         return self.t.render(node=node)